ai_source/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.8
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
MouseInfo==0.1.3
mss==10.0.0
neat-python==0.92
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
six==1.17.0
Werkzeug==3.1.3
